+++ src/Infrastructure/LeanCode.ExternalIdentityProviders/Facebook/Constants.cs
@@ -1,0 +1,7 @@
+namespace LeanCode.ExternalIdentityProviders.Facebook
+{
+    public static class Constants
+    {
+        public const string GrantType = "facebook";
+    }
+}